mka: all engine part refactored - general separation for base  classes: 3 sec compilation improvement, added: template for dynamic two dimensional arrays; interface for scenemanager for adddataobject into visman; some mistakes in core object corrected
git-svn-id: https://localhost:8443/svn/expectance@234 aa805ad0-02f7-422c-84f2-0395ce76bb97

+++ trunk/src/engine/src/core/config.cpp
@@ -204,7 +204,7 @@ int Config::GetIntVal(const char * key)
 float Config::GetFloatVal(const char * key)
 {
     const char * val = GetStrVal(key);
-    return atof (val);
+    return static_cast<float>(atof (val));
 }
 
 /**
